[AN] refactor: deleteBookmark 예외 처리 부분 삭제 #125

+++ android/app/src/main/java/com/onair/hearit/data/BookmarkMapper.kt
@@ -5,8 +5,8 @@ import com.onair.hearit.domain.model.Bookmark
 
 fun BookmarkResponse.Content.toDomain(): Bookmark =
     Bookmark(
-        hearitId = hearitId.toLong(),
-        bookmarkId = bookmarkId.toLong(),
+        hearitId = hearitId,
+        bookmarkId = bookmarkId,
         title = title,
         summary = summary,
         playTime = playTime,
